Created 7 new files and inherited from ViewController.swift for all of them.
The file names are based of of the 7 Big ideas. Also the inheritence from ViewController will not remain the same moving on. This is to just have it there at the start.
NOTE: import Foundation is not needed in the file right now because what we are doing will be handled by UIKit.