Created a new class file called DrawingView and started coding drawings.
Also fixed an error with one of the segues that led to a fatel error with unwrapping a nil due to "showDetail" not being in the identity inspector of one of the segues.

Also added a "cute" trasparent small (20x20 at most) image to the program to be used for later.

NOTE: Code has errors that I will fix.