Unzipped the GameCode file and added them to the Project with new folders.
GameController folder holds the new controllers, sounds is in resources and holds the audio, the Utilities.swift file is simply by itself in resources, starField goes inside the view folder, and 5 game model files goes into the Gamemodel folder.

ALSO there was a huge issue with folders being moved around by accident or copied so there are changes that didn't need to happen with files being deleted and re-added.