Coded the BucketItem file in the model file and created a button to link to the DataViewController screen.
Also changed "randomEmojiSymbol()" to "randomEmoji()" in the BucketItemCell class in order to get rid of an error which also ended up fixing the BucketItem errors that were happening in other sections and files. I am under the assumption that randomEmojiSymbol() was simply not fixed to match the different method name for the random emoji generator.

This finishes the DataViewController section here.